# docs/swagger.yaml
definitions:
...
  controller.PersonalPage:
    properties:
      Songs:
        items:
          $ref: '#/definitions/models.Songs'
        type: array
      admire:
        items:
          $ref: '#/definitions/models.Admire'
        type: array
      avatar:
        type: string
      background:
        type: string
      hide_number:
        type: integer
      more:
        type: string
      name:
        type: string
      phone:
        type: string
      requestSongs:
        items:
          $ref: '#/definitions/models.RequestSongs'
        type: array
      school:
        type: string
      setting1:
        type: integer
      setting2:
        type: integer
      setting3:
        type: integer
      truename:
        type: string
    type: object
  controller.PhoneHealing:
    properties:
      phone:
        type: string
    type: object
  controller.PutUserInf:
    properties:
      avatar:
        type: string
      name:
        type: string
      phone:
        type: string
      setting1:
        type: integer
      setting2:
        type: integer
      setting3:
        type: integer
      signature:
        type: string
      truename:
        type: string
    type: object
  controller.RealResp:
    properties:
      url:
        type: string
    type: object
  controller.RecordParams:
    properties:
      id:
        type: string
      name:
        type: string
      url:
        type: string
    required:
    - id
    - name
    - url
    type: object
  controller.Tag:
    properties:
      tag1:
        type: integer
      tag2:
        type: integer
      tag3:
        type: integer
      tag4:
        type: integer
      tag5:
        type: integer
      tag6:
        type: integer
      tag7:
        type: integer
    type: object
  controller.ToSaveBackground:
    properties:
      background:
        type: integer
    type: object
...
  controller.VodParams:
    properties:
      language:
        type: string
      more:
        type: string
      singer:
        type: string
      songs:
        type: string
      style:
        type: string
    required:
    - language
    - more
    - singer
    - songs
    - style
    type: object
  e.ErrMsgResponse:
    properties:
      message:
        type: string
    type: object
  models.Admire:
    properties:
      from:
        type: string
      id:
        type: integer
      number:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
...
  models.Last:
    properties:
      content:
        type: string
      id:
        type: integer
      stringtime:
        type: string
      time:
        type: string
      type:
        type: integer
      url:
        type: string
    type: object
  models.MainMsg:
    properties:
      listen:
        items:
          $ref: '#/definitions/models.SongMsg'
        type: array
      sing:
        items:
          $ref: '#/definitions/models.SongMsg'
        type: array
    type: object
...
  models.RequestSongs:
    properties:
      hidename:
        type: integer
      id:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
  models.SongMsg:
    properties:
      avatar:
        type: string
      id:
        type: integer
      like:
        type: integer
      more:
        type: string
      name:
        type: string
      senduser:
        type: string
      sex:
        type: integer
      singer:
        type: string
      source:
        type: string
      style:
        type: string
      time:
        type: string
      user:
        type: string
    type: object
  models.Songs:
    properties:
      from:
        type: string
      id:
        type: integer
      song:
        type: string
      time:
        type: string
    type: object
  models.Target:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.ToMessageCell:
    properties:
      content:
        type: string
      fromUserID:
        type: integer
      id:
        type: integer
      stringtime:
        type: string
      time:
        type: string
      toUserID:
        type: integer
      type:
        type: integer
      url:
        type: string
    type: object
  models.ToMessagePage:
    properties:
      last:
        $ref: '#/definitions/models.Last'
      target:
        $ref: '#/definitions/models.Target'
    type: object
...
info:
  contact: {}
  description: 2020治愈系
  title: healing2020
  version: "1.0"
paths:
...
  /api/like:
    get:
      description: 点赞
      parameters:
      - description: type id
        in: query
        name: id
        required: true
        type: string
      - description: 1 song; 2 deliver
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - heal
  /api/main/page:
    get:
      description: 首页数据
      parameters:
      - description: 1综合排序2最新发布
        in: query
        name: sort
        required: true
        type: string
      - description: language
        in: query
        name: language
        type: string
      - description: style
        in: query
        name: style
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MainMsg'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - main
  /api/message:
    get:
      description: 消息首页
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ToMessagePage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - message
  /api/message/cell:
    get:
      description: 用户与另一用户聊天室的具体信息
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ToMessageCell'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - message
...
  /api/record:
    get:
      description: 听录音
      parameters:
      - description: record id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.RealResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - heal
    post:
      description: 录音治愈发布
      parameters:
      - description: 点歌单id
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/controller.RecordParams'
      - description: user name
        in: body
        name: name
        schema:
          $ref: '#/definitions/controller.RecordParams'
      - description: url
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/controller.RecordParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - heal
...
  /api/user:
    get:
      description: 已登录用户的个人页接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PersonalPage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
    put:
      description: 更新用户信息
      parameters:
      - description: 更新的用户信息
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.PutUserInf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
  /api/user/{id}:
    get:
      description: 其它用户的个人页接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PersonalPage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
  /api/user/background:
    put:
      description: 修改个人背景
      parameters:
      - description: 修改后的个人背景
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.ToSaveBackground'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - user
  /api/user/hobby:
    get:
      description: 获取用户爱好
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Tag'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - hobby
    post:
      description: 爱好选择接口
      parameters:
      - description: 用户爱好标签
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - hobby
  /api/user/phone:
    get:
      description: 用户手机
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PhoneHealing'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - heal
...
  /api/vod:
    post:
      description: 点歌
      parameters:
      - description: song's name
        in: body
        name: songs
        required: true
        schema:
          $ref: '#/definitions/controller.VodParams'
      - description: singer
        in: body
        name: singer
        required: true
        schema:
          $ref: '#/definitions/controller.VodParams'
      - description: 备注
        in: body
        name: more
        required: true
        schema:
          $ref: '#/definitions/controller.VodParams'
      - description: style
        in: body
        name: style
        required: true
        schema:
          $ref: '#/definitions/controller.VodParams'
      - description: language
        in: body
        name: language
        required: true
        schema:
          $ref: '#/definitions/controller.VodParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - heal
  /api/vod/hide_name:
    put:
      description: 匿名
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - mypersonalpage
...
  /message:
    post:
      description: 发送消息并保存于数据库
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/e.ErrMsgResponse'
      tags:
      - message
swagger: "2.0"